Move helper methods from public class to util module
