+++ pypgn/game_utils.py
@@ -1,0 +1,38 @@
+import re
+from typing import Mapping, List, NewType, Union
+
+Move = NewType('Move', Union[str, List])
+
+
+def _get_pgn_list(path: str) -> list:
+    with open(file=path, mode='r') as f:
+        lines = f.read().splitlines()
+        return lines
+
+
+def _get_tags(pgn: list) -> Mapping:
+    tag_list = [tag for tag in pgn if re.search(r'^\[', tag)]
+    tag_dict: dict = {}
+    for tag in tag_list:
+        key = tag.split(" ", 1)[0][1:]
+        val = tag.split(" ", 1)[1][1:-2]
+        tag_dict[key] = val
+
+    return tag_dict
+
+
+def _get_moves(pgn: list) -> List[Move]:
+    for line in pgn:
+        if re.search(r'^1\. ', line):
+            movetext: str = line
+
+    movetext_items = movetext.split(" ")
+    moves = []
+    for i, item in enumerate(movetext_items):
+        if re.search(r'\d\.', item):
+            move = [movetext_items[i],
+                    movetext_items[i + 1],
+                    movetext_items[i + 2]]
+            moves.append(move)
+
+    return moves
